// apps/journeys-admin/__generated__/AiGetAdminJourney.ts
export interface AiGetAdminJourney_journey_language {
  __typename: "Language";
  id: string;
  bcp47: string | null;
  iso3: string | null;
  name: AiGetAdminJourney_journey_language_name[];
}
// ...
export type AiGetAdminJourney_journey_blocks = AiGetAdminJourney_journey_blocks_GridContainerBlock | AiGetAdminJourney_journey_blocks_ButtonBlock | AiGetAdminJourney_journey_blocks_CardBlock | AiGetAdminJourney_journey_blocks_IconBlock | AiGetAdminJourney_journey_blocks_ImageBlock | AiGetAdminJourney_journey_blocks_RadioOptionBlock | AiGetAdminJourney_journey_blocks_RadioQuestionBlock | AiGetAdminJourney_journey_blocks_SignUpBlock | AiGetAdminJourney_journey_blocks_SpacerBlock | AiGetAdminJourney_journey_blocks_StepBlock | AiGetAdminJourney_journey_blocks_TextResponseBlock | AiGetAdminJourney_journey_blocks_TypographyBlock | AiGetAdminJourney_journey_blocks_VideoBlock | AiGetAdminJourney_journey_blocks_VideoTriggerBlock;

export interface AiGetAdminJourney_journey_primaryImageBlock {
  __typename: "ImageBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  src: string | null;
  alt: string;
  width: number;
  height: number;
  /**
   * blurhash is a compact representation of a placeholder for an image.
   * Find a frontend implementation at https: // github.com/woltapp/blurhash
   */
  blurhash: string;
  scale: number | null;
  focalTop: number | null;
  focalLeft: number | null;
}

export interface AiGetAdminJourney_journey_creatorImageBlock {
  __typename: "ImageBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  src: string | null;
  alt: string;
  width: number;
  height: number;
  /**
   * blurhash is a compact representation of a placeholder for an image.
   * Find a frontend implementation at https: // github.com/woltapp/blurhash
   */
  blurhash: string;
  scale: number | null;
  focalTop: number | null;
  focalLeft: number | null;
}

export interface AiGetAdminJourney_journey_userJourneys_user {
  __typename: "User";
  id: string;
  firstName: string;
  lastName: string | null;
  imageUrl: string | null;
}

export interface AiGetAdminJourney_journey_userJourneys {
  __typename: "UserJourney";
  id: string;
  role: UserJourneyRole;
  /**
   * Date time of when the journey was first opened
   */
  openedAt: any | null;
  user: AiGetAdminJourney_journey_userJourneys_user | null;
}

export interface AiGetAdminJourney_journey_chatButtons {
  __typename: "ChatButton";
  id: string;
  link: string | null;
  platform: MessagePlatform | null;
}

export interface AiGetAdminJourney_journey_host {
  __typename: "Host";
  id: string;
  teamId: string;
  title: string;
  location: string | null;
  src1: string | null;
  src2: string | null;
}

export interface AiGetAdminJourney_journey_team {
  __typename: "Team";
  id: string;
  title: string;
  publicTitle: string | null;
}

export interface AiGetAdminJourney_journey_tags_name_language {
  __typename: "Language";
  id: string;
}

export interface AiGetAdminJourney_journey_tags_name {
  __typename: "TagName";
  value: string;
  language: AiGetAdminJourney_journey_tags_name_language;
  primary: boolean;
}

export interface AiGetAdminJourney_journey_tags {
  __typename: "Tag";
  id: string;
  parentId: string | null;
  name: AiGetAdminJourney_journey_tags_name[];
}

export interface AiGetAdminJourney_journey_logoImageBlock {
  __typename: "ImageBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  src: string | null;
  alt: string;
  width: number;
  height: number;
  /**
   * blurhash is a compact representation of a placeholder for an image.
   * Find a frontend implementation at https: // github.com/woltapp/blurhash
   */
  blurhash: string;
  scale: number | null;
  focalTop: number | null;
  focalLeft: number | null;
}

export interface AiGetAdminJourney_journey_menuStepBlock {
  __typename: "StepBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  /**
   * locked will be set to true if the user should not be able to manually
   * advance to the next step.
   */
  locked: boolean;
  /**
   * nextBlockId contains the preferred block to navigate to, users will have to
   * manually set the next block they want to card to navigate to
   */
  nextBlockId: string | null;
  /**
   * Slug should be unique amongst all blocks
   * (server will throw BAD_USER_INPUT error if not)
   * If not required will use the current block id
   * If the generated slug is not unique the uuid will be placed
   * at the end of the slug guaranteeing uniqueness
   */
  slug: string | null;
}

export interface AiGetAdminJourney_journey {
  __typename: "Journey";
  id: string;
  slug: string;
  /**
   * private title for creators
   */
  title: string;
  description: string | null;
  status: JourneyStatus;
  language: AiGetAdminJourney_journey_language;
  createdAt: any;
  featuredAt: any | null;
  publishedAt: any | null;
  themeName: ThemeName;
  themeMode: ThemeMode;
  strategySlug: string | null;
  /**
   * title for seo and sharing
   */
  seoTitle: string | null;
  seoDescription: string | null;
  template: boolean | null;
  blocks: AiGetAdminJourney_journey_blocks[] | null;
  primaryImageBlock: AiGetAdminJourney_journey_primaryImageBlock | null;
  creatorDescription: string | null;
  creatorImageBlock: AiGetAdminJourney_journey_creatorImageBlock | null;
  userJourneys: AiGetAdminJourney_journey_userJourneys[] | null;
  chatButtons: AiGetAdminJourney_journey_chatButtons[];
  host: AiGetAdminJourney_journey_host | null;
  team: AiGetAdminJourney_journey_team | null;
  tags: AiGetAdminJourney_journey_tags[];
  website: boolean | null;
  showShareButton: boolean | null;
  showLikeButton: boolean | null;
  showDislikeButton: boolean | null;
  /**
   * public title for viewers
   */
  displayTitle: string | null;
  logoImageBlock: AiGetAdminJourney_journey_logoImageBlock | null;
  menuButtonIcon: JourneyMenuButtonIcon | null;
  menuStepBlock: AiGetAdminJourney_journey_menuStepBlock | null;
}

export interface AiGetAdminJourney {
  journey: AiGetAdminJourney_journey;
}

export interface AiGetAdminJourneyVariables {
  id: string;
}